requirements.txt
# Core dependencies for LifeOS Autonomous System
aiohttp>=3.9.0
python-dotenv>=1.0.0
loguru>=0.7.0
flask>=2.3.0
psutil>=5.9.0

# Optional dependencies for enhanced features
notion-client>=2.2.1
google-api-python-client>=2.119.0
google-generativeai>=0.3.2
youtube-transcript-api>=0.6.2
yt-dlp>=2024.3.10

# Development dependencies
requests>=2.31.0
asyncio-throttle>=1.0.2
watchfiles>=0.21.0